Refactor Prompts with PromptLayout
